src/lucy/requirements.txt
discord.py
emoji
gradle_client
pip
pubchempy
pillow
pyyaml
requests
rdkit
selenium
webdriver_manager
